Deletes app module and creates presentation module
