+++ presentation/src/main/res/mipmap-anydpi-v33/ic_launcher.xml
